v1.0.0.35 - manuelles Nachstempeln implementiert - verschiedene Fehlerkorrekturen - Warnungen bearbeitet und geklärt - Versionsupdate bei Imports - Designanpassung für DarkMode - keine Stempelung ohne usernamen in den Settings - Datenbank angepasst - PendingEvents nun im Klartext und jedes Event kann einzeln gelöscht werden

+++ app/src/main/java/org/fx/timetracker/TimeEvent.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "pending_events")
 data class TimeEvent(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0L,
     val jsonPayload: String // Wir speichern einfach den kompletten JSON-String
 )
